change numeric data

+++ BinDec.c
@@ -5,7 +5,7 @@
 int main(void) // void main()
 {
 	int i;
-	long input;
+	short input;
 	
 	printf("10진수 -> 2진수 변환기\n");
 	printf("0을 입력시 프로그램이 종료됩니다.\n");
@@ -15,7 +15,7 @@ int main(void) // void main()
 		printf("정수를 입력하세요 : ");
 		scanf("%d", &input);
 		
-		if (input == 0)
+		if(input == 0)
 			break;
 		printf("2진수 : ");
 		for(i = 0; i < 16; i++)
